@N Merged SYNTAX_1_3_BRANCH into HEAD

+++ src/de/willuhn/jameica/fibu/gui/action/KontoDelete.java
@@ -1,7 +1,7 @@
 /**********************************************************************
  * $Source: /cvsroot/syntax/syntax/src/de/willuhn/jameica/fibu/gui/action/KontoDelete.java,v $
- * $Revision: 1.8 $
- * $Date: 2008/02/26 19:13:23 $
+ * $Revision: 1.9 $
+ * $Date: 2009/07/03 10:52:19 $
  * $Author: willuhn $
  * $Locker:  $
  * $State: Exp $
@@ -46,6 +46,8 @@ public class KontoDelete implements Action
       if (context instanceof Konto)
       {
         k = new Konto[] {(Konto) context};
+        if (!k[0].isUserObject())
+          throw new ApplicationException(i18n.tr("System-Konten dürfen nicht gelöscht werden"));
       }
       else
         k = (Konto[]) context;
@@ -76,9 +78,12 @@ public class KontoDelete implements Action
         String s = null;
         for (int i=0;i<k.length;++i)
         {
-          s = k[i].getKontonummer();
-          k[i].delete();
-          deleted++;
+          if (k[i].isUserObject())
+          {
+            s = k[i].getKontonummer();
+            k[i].delete();
+            deleted++;
+          }
         }
         k[0].transactionCommit();
 
@@ -116,8 +121,8 @@ public class KontoDelete implements Action
 
 /*********************************************************************
  * $Log: KontoDelete.java,v $
- * Revision 1.8  2008/02/26 19:13:23  willuhn
- * *** empty log message ***
+ * Revision 1.9  2009/07/03 10:52:19  willuhn
+ * @N Merged SYNTAX_1_3_BRANCH into HEAD
  *
  * Revision 1.7  2006/01/02 15:18:29  willuhn
  * @N Buchungs-Vorlagen
